+++ lib/src/content/component_name.dart
@@ -1,11 +1,12 @@
 /* This is free and unencumbered software released into the public domain. */
 
-import '../os/parcelable.dart';
+import '../os/parcel.dart' show Parcel;
+import '../os/parcelable.dart' show Parcelable;
 
 /// Identifier for a specific application component that is available.
 ///
 /// See: https://developer.android.com/reference/android/content/ComponentName
-class ComponentName implements Parcelable {
+class ComponentName with Parcelable {
   /// The package name of this component.
   ///
   /// See: https://developer.android.com/reference/android/content/ComponentName#getPackageName()
@@ -29,5 +30,13 @@ class ComponentName implements Parcelable {
   /// See: https://developer.android.com/reference/android/content/ComponentName#flattenToString()
   String flattenToString() {
     return <String>[packageName, className].join("/");
+  }
+
+  @override
+  String get parcelableCreator => "android.content.ComponentName";
+
+  @override
+  void writeToParcel(final Parcel parcel, [final int flags = 0]) {
+    throw UnimplementedError(); // TODO
   }
 }